Ignored key files and changed data model
In my first commit, I accidentally included a sqlite db and the settings.py
file, which includes a Django secret_key. So those have been removed, and ignored
in the repo for this commit. I'm messing around with admin templates, but the two
templates included in this commit don't actually do anything.

Added a hotel promo code line to the data model for Camps.